Delete BSI directory
HI Arun, I hope this message finds you well. My name is Pasya, If there's a chance, I would love to have a conversation with you about this repository,
 
I regret making this repository public and accessible to others. May I request that it be deleted please ? 

I will provide a better version of the source code later. The main reason is that it contains the name of a company, which I believe is not ethically appropriate to publish. I apologize for any inconvenience this may have caused. 

Thank You in Advance